+++ tools/oicana_cli/src/compile.rs
@@ -88,7 +88,7 @@ pub fn compile(args: CompileArgs) -> anyhow::Result<()> {
         None => Path::new("."),
         Some(ref template) => Path::new(template),
     };
-    let mut template = Template::<NativeTemplate>::init_with_fonts(path, &args.fonts.load())?;
+    let mut template = Template::open_with_fonts(path, &args.fonts.load())?;
     template.set_diagnostic_color(diagnostic_color());
     let name: String = template.manifest().package.name.to_string();
     info!("Compiling template '{name}'.");